+++ rich/table.py
@@ -315,7 +315,7 @@ class Table:
                 render_text = text
             else:
                 render_text = console.render_str(text, style=style)
-            return render_text.wrap(table_width, "center")
+            return render_text.wrap(console, table_width, "center")
 
         if self.title:
             yield render_annotation(