+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2020-05-10
+
+### Added
+
+- Added hyperlinks to Style and markup
+- Added justify and code theme switches to markdown command
+
 ## [1.0.3] - 2020-05-08
 
 ### Added